Revisions to 1st and 2nd level
1st Level: Increase the timer for the laser field over the goo. Modified
room with the switch to that laser field so the player cannot get
trapped there.
2nd Level: Lowered some roofs. Altered some lighting. Added lights near
the entrance to the third floor to guide the player there. Moved the
cube in the fourth floor backroom so it is visible when the player
stands on the regular button. Made the backroom smaller. Added goo
between the fourth and fifth floors.