Iteration 3 of Levels 1 and 2
Fixed Level 1 - Removed the arbitrary laser - made it so you cannot skip
the lightbrigde.
Changed Floor 3 of Level 2 to make it more interesting. Added in a
sphere button to open the glass. Added a sphere button on Floor 4 to
turn off a laser field (blocking Floor 5). Added the sphere cube to
Floor 3.5 behind an emancipation field that can be turned off with a
laser catcher. Also opened up Floor 4 to allow quicker access to Floors
1 and 2 rather than having to walk all the way around.